% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{annot_g_synechocystis_6803}
\alias{annot_g_synechocystis_6803}
\title{A data frame corresponding to a gff file for Synechocystis 6803 example data
A list containing all necessary information from a gff file for visualization
and final output.}
\format{
A list with 2 items:
\describe{
\item{data annotation:}{a data frame with 5853 rows and 6 variables:
\describe{
\item{region:}{the region from the gff file}
\item{start:}{the start of the annotation}
\item{end:}{the end of the annotation}
\item{strand:}{the strand of the annotation}
\item{gene:}{the annotated gene name}
\item{locus_tag:}{the annotated locus tag}
}
}
\item{genome length:}{a numeric vector containing the length of the genome}
}
}
\source{
\url{https://github.com/CyanolabFreiburg/Transcriptome_wide_decay}
}
\usage{
data(annot_g_synechocystis_6803)
}
\description{
A data frame corresponding to a gff file for Synechocystis 6803 example data
A list containing all necessary information from a gff file for visualization
and final output.
}
\keyword{datasets}
