% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{example_input_minimal}
\alias{example_input_minimal}
\title{An artificial example SummarizedExperiment
An example SummarizedExperiment containing information about the intensities
at all time points (assay). Seqnames, IRanges and strand columns (rowRanges)
and colData with time point series and replicates.}
\format{
An object of class \code{RangedSummarizedExperiment} with 4 rows and 33 columns.
}
\source{
\url{https://github.com/CyanolabFreiburg/rifi}
}
\usage{
data(example_input_minimal)
}
\description{
An artificial example SummarizedExperiment
An example SummarizedExperiment containing information about the intensities
at all time points (assay). Seqnames, IRanges and strand columns (rowRanges)
and colData with time point series and replicates.
}
\keyword{datasets}
