% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{example_input_synechocystis_6803}
\alias{example_input_synechocystis_6803}
\title{An example input data frame from Synechocystis PCC 6803.#'
A data set from microarrays data containing information about the intensities
at all time points, as well as position, strand and ID information.}
\format{
A data frame with 3000 rows and 10 variables:#'
\describe{
\item{0:}{relative intensities at 0 min}
\item{2:}{relative intensities at 2 min}
\item{4:}{relative intensities at 4 min}
\item{8:}{relative intensities at 8 min}
\item{16:}{relative intensities at 16 min}
\item{32:}{relative intensities at 32 min}
\item{64:}{relative intensities at 64 min}
\item{ID:}{unique IDs}
\item{position:}{genome positions}
\item{strand:}{strand information}
}
}
\source{
\url{https://github.com/CyanolabFreiburg/Transcriptome_wide_decay}
}
\usage{
data(example_input_synechocystis_6803)
}
\description{
An example input data frame from Synechocystis PCC 6803.#'
A data set from microarrays data containing information about the intensities
at all time points, as well as position, strand and ID information.
}
\keyword{datasets}
