% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_PDD.r
\name{finding_PDD}
\alias{finding_PDD}
\title{finding_PDD: flags potential candidates for post transcription decay.
'finding_PDD' uses 'score_fun_linear_PDD' to make groups by the difference
to the slope.
Then the slope is checked for steepness to decide for PDD.
'\emph{PDD}' is added to the 'flag' column.
Post transcription decay is characterized by a strong decrease of intensity
by position.
The data frame needs to contain at least 'ID', 'intensity', 'position' and
'position_segment'!}
\usage{
finding_PDD(probe, cores = 1, pen = 2, pen_out = 1, thrsh = 0.001)
}
\arguments{
\item{probe}{data frame: the probe based data frame.}

\item{cores}{interger: the number of assigned cores for the task}

\item{pen}{numeric: an internal parameter for the dynamic programming.
Higher values result in fewer fragments. Advised to be kept at 2.
Default is 2.}

\item{pen_out}{numeric: an internal parameter for the dynamic programming.
Higher values result in fewer possible outliers. Advised to be kept at 1.
Default is 1.}

\item{thrsh}{numeric: an internal parameter that allows fragments with slopes
steeper than the thrsh to be flagged with '\emph{PDD}'. Higher values result in
fewer candidates. Advised to be kept at 0.001. Default is 0.001.}
}
\value{
the probe based data frame with the modified flag:
\describe{
\item{ID:}{The bin/probe specific ID}
\item{position:}{The bin/probe specific position}
\item{strand:}{The bin/probe specific strand}
\item{intensity:}{The relative intensity at time point 0}
\item{probe_TI:}{An internal value to determine which fitting model is
applied}
\item{flag:}{Information on which fitting model is applied}
\item{postion_segment:}{The position based segment}
}
}
\description{
finding_PDD: flags potential candidates for post transcription decay.
'finding_PDD' uses 'score_fun_linear_PDD' to make groups by the difference
to the slope.
Then the slope is checked for steepness to decide for PDD.
'\emph{PDD}' is added to the 'flag' column.
Post transcription decay is characterized by a strong decrease of intensity
by position.
The data frame needs to contain at least 'ID', 'intensity', 'position' and
'position_segment'!
}
\examples{
data(preprocess_minimal)
finding_PDD(probe = preprocess_minimal$probe_df,  cores = 2, pen = 2,
pen_out = 1, thrsh = 0.001)

}
