% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_TI.r
\name{finding_TI}
\alias{finding_TI}
\title{finding_TI: flags potential candidates for transcription interference.
'finding_TI' uses 'score_fun_ave' to make groups by the mean of "probe_TI".
"TI" is added to the "flag" column.
TI is characterized by relative intensities at time points later than "0".
The data frame needs to contain at least "ID", "probe_TI" and
"position_segment"!}
\usage{
finding_TI(probe, cores, pen = 10, thrsh = 0.5, add = 1000)
}
\arguments{
\item{probe}{data frame: the probe based data frame.}

\item{cores}{integer: the number of assigned cores for the task}

\item{pen}{numeric: an internal parameter for the dynamic programming.
Higher values result in fewer fragments. Advised to be kept at 10.
Default is 10.}

\item{thrsh}{numeric: an internal parameter that allows fragments with a
certain amount of IDs with higher relative intensities at time points later
than "0" to be flagged as "\emph{TI}". Higher values result in fewer candidates.
-0.5 is 25 \%, 0 is 50\%, 0.5 is 75\%. Advised to be kept at 0.5.
Default is 0.5.}

\item{add}{integer: range of nucleotides before and after a potential TI
event wherein IDs are fitted with the TI fit.}
}
\value{
the probe based data frame with the modified flag:
\describe{
\item{ID:}{The bin/probe specific ID}
\item{position:}{The bin/probe specific position}
\item{strand:}{The bin/probe specific strand}
\item{intensity:}{The relative intensity at time point 0}
\item{probe_TI:}{An internal value to determine which fitting model is
applied}
\item{flag:}{Information on which fitting model is applied}
\item{postion_segment:}{The position based segment}
}
}
\description{
finding_TI: flags potential candidates for transcription interference.
'finding_TI' uses 'score_fun_ave' to make groups by the mean of "probe_TI".
"TI" is added to the "flag" column.
TI is characterized by relative intensities at time points later than "0".
The data frame needs to contain at least "ID", "probe_TI" and
"position_segment"!
}
\examples{
data(preprocess_e_coli)
finding_TI(probe = preprocess_e_coli$probe_df, cores = 2, pen = 10,
thrsh = 0.5, add = 1000)

}
