% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{fit_minimal}
\alias{fit_minimal}
\title{The artificial result of rifi_fit for artificial example data
A data frame containing the output from rifi_fit.}
\format{
A data frame with 24 rows and 10 variables:
\describe{
\item{ID:}{The bin/probe specific ID}
\item{position:}{The bin/probe specific position}
\item{strand:}{The bin/probe specific strand}
\item{intensity:}{The relative intensity at time point 0}
\item{probe_TI:}{An internal value to determine which fitting model
is applied}
\item{flag:}{Information on which fitting model is applied}
\item{position_segment:}{The position based segment}
\item{delay:}{The delay value of the bin/probe}
\item{half_life:}{The half-life of the bin/probe}
\item{TI_termination_factor:}{The termination factor of the bin/probe}
}
}
\source{
\url{https://github.com/CyanolabFreiburg/Transcriptome_wide_decay}
}
\usage{
data(fit_minimal)
}
\description{
The artificial result of rifi_fit for artificial example data
A data frame containing the output from rifi_fit.
}
\keyword{datasets}
