% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragment_inty.r
\name{fragment_inty}
\alias{fragment_inty}
\title{\subsection{=========================================================================
fragment_inty}{

fragment_inty performs the intensity fragmentation
}}
\usage{
fragment_inty(inp, cores = 1, pen, pen_out)
}
\arguments{
\item{inp}{SummarizedExperiment: the input data frame with correct format.}

\item{cores}{cores: integer: the number of assigned cores for the task.}

\item{pen}{numeric: an internal parameter for the dynamic programming.
Higher values result in fewer fragments. Default is the auto generated value.}

\item{pen_out}{numeric: an internal parameter for the dynamic programming.
Higher values result in fewer allowed outliers. Default is the auto generated
value.}
}
\value{
the SummarizedExperiment object: with intensity_fragment and
intensity_mean_fragment added to the rowRanges.
}
\description{
fragment_inty makes intensity_fragments based on HL_fragments and assigns
all gathered information to the SummarizedExperiment object.
}
\details{
The columns "intensity_fragment" and "intensity_mean_fragment" are added.

fragment_inty makes intensity_fragments and assigns the mean of each
fragment.

The function used is:

.score_fun_ave.

The input is the the SummarizedExperiment object.

pen is the penalty for new fragments in the dynamic programming, pen_out
is the outlier penalty.
}
\examples{
data(fragmentation_minimal)
fragment_inty(inp = fragmentation_minimal, cores = 2, pen = 2, pen_out = 1)

}
