% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{fragmentation_minimal}
\alias{fragmentation_minimal}
\title{The result of rifi_fragmentation for artificial example data
A data frame containing the output from rifi_fragmentation, the probe_df}
\format{
An object of class \code{RangedSummarizedExperiment} with 24 rows and 33 columns.
}
\source{
\url{https://github.com/CyanolabFreiburg/Transcriptome_wide_decay}
}
\usage{
data(fragmentation_minimal)
}
\description{
The result of rifi_fragmentation for artificial example data
A data frame containing the output from rifi_fragmentation, the probe_df
}
\keyword{datasets}
