% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{fragmentation_synechocystis_6803}
\alias{fragmentation_synechocystis_6803}
\title{The result of rifi_fragmentation for Synechocystis 6803 example data
A SummarizedExperiment containing the output from rifi_fragmentation as an
extension fo rowRanges}
\format{
rowRanges of the SummarizedExperiment:
\describe{
\item{ID:}{The bin/probe specific ID}
\item{position:}{The bin/probe specific position}
\item{intensity:}{The relative intensity at time point 0}
\item{probe_TI:}{An internal value to determine which fitting model is
applied}
\item{flag:}{Information on which fitting model is applied}
\item{position_segment:}{The position based segment}
\item{delay:}{The delay value of the bin/probe}
\item{half_life:}{The half-life of the bin/probe}
\item{TI_termination_factor:}{The termination factor of the bin/probe}
\item{delay_fragment:}{The delay fragment the bin belongs to}
\item{velocity_fragment:}{The velocity value of the respective delay
fragment}
\item{intercept:}{The vintercept of fit through the respective delay
fragment}
\item{slope:}{The slope of the fit through the respective delay fragment}
\item{HL_fragment:}{The half-life fragment the bin belongs to}
\item{HL_mean_fragment:}{The mean half-life value of the respective
half-life fragment}
\item{intensity_fragment:}{The intensity fragment the bin belongs to}
\item{intensity_mean_fragment:}{The mean intensity value of the
respective intensity fragment}
\item{TU:}{The overarching transcription unit}
\item{TI_termination_fragment:}{The TI fragment the bin belongs to}
\item{TI_mean_termination_factor:}{The mean termination factor of the
respective TI fragment}
\item{seg_ID:}{The combined ID of the fragment}
}
}
\source{
\url{https://github.com/CyanolabFreiburg/rifi}
}
\usage{
data(fragmentation_synechocystis_6803)
}
\description{
The result of rifi_fragmentation for Synechocystis 6803 example data
A SummarizedExperiment containing the output from rifi_fragmentation as an
extension fo rowRanges
}
\keyword{datasets}
