% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_df.r
\name{make_df}
\alias{make_df}
\title{make_df: converts the input data frame to the template probe based dataframe.
'make_df' creates the probe based data frame with the columns: "ID",
"position","strand","intensity","probe_TI" and "flag" from the input
dataframe.
The replicates are collapsed into their respective means.
"ID", "position" and "strand" are taken over as is.
"intensity" is the mean intensity from time point 0.
"probe_TI" is a value needed for the distribution for the different fitting
models.
"flag" contains information or the distribution for the different fitting
models.
Here, probes that don't reach the background level expression are flagged
as "\emph{ABG}" ("above background").
This is only needed for microarray data and is controlled by the bg
parameter.
The default for bg = 0, resulting in all probes to be above background
(0 is advised for RNAseq data).
Probes where all replicates were filtered in the optional filtration step
can be fully removed by rm_FLT = TRUE!}
\usage{
make_df(inp, cores = 1, bg = 0, rm_FLT = FALSE)
}
\arguments{
\item{inp}{data frame: the (checked) input data frame.}

\item{cores}{integer: the number of assigned cores for the task.}

\item{bg}{numeric: threshold over which the last timepoint has to be fitted
with the above background mode.}

\item{rm_FLT}{logical: remove IDs where all replicates are marked as
filtered. Default is FALSE.}
}
\value{
the probe based data frame:
\describe{
\item{ID:}{The bin/probe specific ID}
\item{position:}{The bin/probe specific position}
\item{strand:}{The bin/probe specific strand}
\item{intensity:}{The relative intensity at time point 0}
\item{probe_TI:}{An internal value to determine which fitting model is
applied}
\item{flag:}{Information on which fitting model is applied}
}
}
\description{
make_df: converts the input data frame to the template probe based dataframe.
'make_df' creates the probe based data frame with the columns: "ID",
"position","strand","intensity","probe_TI" and "flag" from the input
dataframe.
The replicates are collapsed into their respective means.
"ID", "position" and "strand" are taken over as is.
"intensity" is the mean intensity from time point 0.
"probe_TI" is a value needed for the distribution for the different fitting
models.
"flag" contains information or the distribution for the different fitting
models.
Here, probes that don't reach the background level expression are flagged
as "\emph{ABG}" ("above background").
This is only needed for microarray data and is controlled by the bg
parameter.
The default for bg = 0, resulting in all probes to be above background
(0 is advised for RNAseq data).
Probes where all replicates were filtered in the optional filtration step
can be fully removed by rm_FLT = TRUE!
}
\examples{
data(preprocess_minimal)
make_df(inp = preprocess_minimal$input_df, cores = 2, bg = 0,
rm_FLT = FALSE)

}
