% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{penalties_e_coli}
\alias{penalties_e_coli}
\title{The result of rifi_penalties for E.coli example data.
A SummarizedExperiment containing the output from rifi_penalties including
the logbook and the four penalty objects as metadata.}
\format{
A list with 5 items:
\describe{
\item{logbook:}{The logbook vector containing all penalty information}
\item{pen_obj_delay:}{A list with 4 items:
\describe{
\item{logbook:}{The logbook vector containing all penalty information}
\item{delay_penalties:}{a vetor with the delay penalty and delay
outlier penalty}
\item{correct:}{a matrix of the correct splits}
\item{wrong:}{a matrix of the incorrect splits}
}
}
\item{pen_obj_HL:}{A list with 4 items:
\describe{
\item{logbook:}{The logbook vector containing all penalty information}
\item{HL_penalties:}{a vetor with the half-life penalty and half-life
outlier penalty}
\item{correct:}{a matrix of the correct splits}
\item{wrong:}{a matrix of the incorrect splits}
}
}
\item{pen_obj_inty:}{A list with 4 items:
\describe{
\item{logbook:}{The logbook vector containing all penalty information}
\item{inty_penalties:}{a vetor with the intensity penalty and intensity
outlier penalty}
\item{correct:}{a matrix of the correct splits}
\item{wrong:}{a matrix of the incorrect splits}
}
}
\item{pen_obj_TI:}{A list with 4 items:
\describe{
\item{logbook:}{The logbook vector containing all penalty information}
\item{TI_penalties:}{a vetor with the TI penalty and TI outlier
penalty}
\item{correct:}{a matrix of the correct splits}
\item{wrong:}{a matrix of the incorrect splits}
}
}
}
}
\source{
\url{https://github.com/CyanolabFreiburg/rifi}
}
\usage{
data(penalties_e_coli)
}
\description{
The result of rifi_penalties for E.coli example data.
A SummarizedExperiment containing the output from rifi_penalties including
the logbook and the four penalty objects as metadata.
}
\keyword{datasets}
