% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{penalties_minimal}
\alias{penalties_minimal}
\title{The result of rifi_penalties for artificial example data
A SummarizedExperiment containing the output from rifi_penalties including
the logbook and the four penalty objects as metadata.}
\format{
An object of class \code{RangedSummarizedExperiment} with 24 rows and 33 columns.
}
\source{
\url{https://github.com/CyanolabFreiburg/rifi}
}
\usage{
data(penalties_minimal)
}
\description{
The result of rifi_penalties for artificial example data
A SummarizedExperiment containing the output from rifi_penalties including
the logbook and the four penalty objects as metadata.
}
\keyword{datasets}
