% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{res_minimal}
\alias{res_minimal}
\title{The result of event_dataframe for E.coli artificial example.
A data frame combining the processed genome annotation and a
summarizedExperiment data from rifi_stats. The dataframe is}
\format{
A list with 2 items:
\describe{
\item{region:}{the region from the gff file}
\item{gene:}{the annotated gene name}
\item{locus_tag:}{the annotated locus tag}
\item{strand:}{the strand of the annotation}
\item{TU:}{The overarching transcription unit}
\item{position:}{The bin/probe specific position}
\item{FC_fragment_intensity:}{}
\item{FC_intensity:}{}
\item{p_value_intensity:}{}
\item{FC_fragment_HL:}{}
\item{FC_HL:}{}
\item{p_value_HL:}{}
\item{FC_HL_intensity_fragment:}{}
\item{FC_HL_intensity:}{}
\item{FC_HL_adapted:}{}
\item{p_value_Manova:}{}
\item{synthesis_ratio:}{}
\item{synthesis_ratio_event:}{}
\item{pausing_site:}{}
\item{iTSS_I:}{}
\item{event_ps_itss_p_value_Ttest:}{}
\item{ps_ts_fragment:}{}
\item{event_position:}{}
\item{event_duration:}{}
\item{delay_frg_slope:}{}
\item{p_value_slope:}{}
\item{delay:}{}
\item{half_life:}{}
\item{intensity:}{}
}
}
\source{
\url{https://github.com/CyanolabFreiburg/rifi}
}
\usage{
data(res_minimal)
}
\description{
The result of event_dataframe for E.coli artificial example.
A data frame combining the processed genome annotation and a
summarizedExperiment data from rifi_stats. The dataframe is
}
\keyword{datasets}
