% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rifi_summary.r
\name{rifi_summary}
\alias{rifi_summary}
\title{rifi_summary: conveniently wraps and summarize all rifi outputs.
Wraps the functions: event_dataframe, dataframe_summary,
dataframe_summary_events, dataframe_summary_events_HL_int,
dataframe_summary_events_ps_itss, dataframe_summary_events_velocity and
dataframe_summary_TI.}
\usage{
rifi_summary(probe, data_annotation)
}
\arguments{
\item{probe}{data frame: the probe based data frame.}

\item{data_annotation}{dataframe: the coordinates are extracted from gff3.}
}
\value{
WIP
}
\description{
rifi_summary: conveniently wraps and summarize all rifi outputs.
Wraps the functions: event_dataframe, dataframe_summary,
dataframe_summary_events, dataframe_summary_events_HL_int,
dataframe_summary_events_ps_itss, dataframe_summary_events_velocity and
dataframe_summary_TI.
}
\examples{
data(stats_minimal)
data(annot_g_minimal)
rifi_summary(probe = stats_minimal, data_annotation =
annot_g_minimal[[1]])

}
\seealso{
\code{event_dataframe}

\code{dataframe_summary}

\code{dataframe_summary_events}

\code{dataframe_summary_events_HL_int}

\code{dataframe_summary_events_ps_itss}

\code{dataframe_summary_events_velocity}

\code{dataframe_summary_TI}
}
