% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_pos.r
\name{segment_pos}
\alias{segment_pos}
\title{segment_pos: divides all IDs by position into position_segments.
segment_pos adds the column "position_segment" to the probe based dataframe.
To reduce run time, the data is divided by regions of no expression larger
than "dist" nucleotides.
The data frame needs to contain at least "strand" and "position".}
\usage{
segment_pos(probe, dista = 300)
}
\arguments{
\item{probe}{data frame: the probe based data frame.}

\item{dista}{integer: the amount of nucleotides defining the gap. Default
is 300.}
}
\value{
the probe based data frame with the position_segment column added:
\describe{
\item{ID:}{The bin/probe specific ID}
\item{position:}{The bin/probe specific position}
\item{strand:}{The bin/probe specific strand}
\item{intensity:}{The relative intensity at time point 0}
\item{probe_TI:}{An internal value to determine which fitting model is
applied}
\item{flag:}{Information on which fitting model is applied}
\item{postion_segment:}{The position based segment}
}
}
\description{
segment_pos: divides all IDs by position into position_segments.
segment_pos adds the column "position_segment" to the probe based dataframe.
To reduce run time, the data is divided by regions of no expression larger
than "dist" nucleotides.
The data frame needs to contain at least "strand" and "position".
}
\examples{
data(preprocess_minimal)
segment_pos(probe = preprocess_minimal$probe_df, dista = 300)

}
