% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{stats_minimal}
\alias{stats_minimal}
\title{The result of rifi_stats for artificial example data
A SummarizedExperiment containing the output of rifi_stats as an
extention to rowRanges and metadata (gff file processed, see gff file
documentation)}
\format{
A rowRanges of SummarizedExperiment with 24 rows and 45 variables:
\describe{
\item{ID:}{The bin/probe specific ID}
\item{position:}{The bin/probe specific position}
\item{intensity:}{The relative intensity at time point 0}
\item{probe_TI:}{An internal value to determine which fitting model is
applied}
\item{flag:}{Information on which fitting model is applied}
\item{position_segment:}{The position based segment}
\item{delay:}{The delay value of the bin/probe}
\item{half_life:}{The half-life of the bin/probe}
\item{TI_termination_factor:}{The termination factor of the bin/probe}
\item{delay_fragment:}{The delay fragment the bin belongs to}
\item{velocity_fragment:}{The velocity value of the respective delay
fragment}
\item{intercept:}{The vintercept of fit through the respective delay
fragment}
\item{slope:}{The slope of the fit through the respective delay fragment}
\item{HL_fragment:}{The half-life fragment the bin belongs to}
\item{HL_mean_fragment:}{The mean half-life value of the respective
half-life fragment}
\item{intensity_fragment:}{The intensity fragment the bin belongs to}
\item{intensity_mean_fragment:}{The mean intensity value of the respective
intensity fragment}
\item{TU:}{The overarching transcription unit}
\item{TI_termination_fragment:}{The TI fragment the bin belongs to}
\item{TI_mean_termination_factor:}{The mean termination factor of the
respective TI fragment}
\item{seg_ID:}{The combined ID of the fragment}
\item{pausing_site:}{}
\item{iTSS_I:}{}
\item{ps_ts_fragment:}{}
\item{event_ps_itss_p_value_Ttest:}{}
\item{p_value_slope:}{}
\item{delay_frg_slope:}{}
\item{velocity_ratio:}{}
\item{event_duration:}{}
\item{event_position:}{}
\item{FC_HL:}{}
\item{FC_fragment_HL:}{}
\item{p_value_HL:}{}
\item{FC_intensity:}{}
\item{FC_fragment_intensity:}{}
\item{p_value_intensity:}{}
\item{FC_HL_intensity:}{}
\item{FC_HL_intensity_fragment:}{}
\item{FC_HL_adapted:}{}
\item{synthesis_ratio:}{}
\item{synthesis_ratio_event:}{}
\item{p_value_Manova:}{}
\item{p_value_TI:}{}
\item{TI_fragments_p_value:}{}
}
}
\source{
\url{https://github.com/CyanolabFreiburg/rifi}
}
\usage{
data(stats_minimal)
}
\description{
The result of rifi_stats for artificial example data
A SummarizedExperiment containing the output of rifi_stats as an
extention to rowRanges and metadata (gff file processed, see gff file
documentation)
}
\keyword{datasets}
