% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{summary_e_coli}
\alias{summary_e_coli}
\title{The result of rifi_summary for E.coli example data
A SummarizedExperiment containing the output of rifi_stats as an
extention to rowRanges}
\format{
The rowRanges of SummarizedExperiment:
\describe{
\item{bin_df:}{all information regarding bins:
\describe{
\item{feature_type:}{}
\item{gene:}{}
\item{locus_tag:}{}
\item{TU:}{The overarching transcription unit}
\item{delay_fragment:}{The delay fragment the bin belongs to}
\item{HL_fragment:}{The half-life fragment the bin belongs to}
\item{half_life:}{The half-life of the bin/probe}
\item{intensity_fragment:}{The intensity fragment the bin belongs to}
\item{intensity:}{The relative intensity at time point 0}
\item{velocity:}{The velocity value of the bin}
}
}
\item{frag_df:}{all information regarding fragments:
\describe{
\item{feature_type:}{}
\item{gene:}{}
\item{locus_tag:}{}
\item{strand:}{The bin/probe specific strand}
\item{TU:}{The overarching transcription unit}
\item{delay_fragment:}{The delay fragment the bin belongs to}
\item{HL_fragment:}{The half-life fragment the bin belongs to}
\item{half_life:}{The half-life of the fragment}
\item{intensity_fragment:}{The intensity fragment the bin belongs to}
\item{intensity:}{The relative intensity at time point 0}
\item{velocity:}{The velocity value of the respective delay fragment}
}
}
\item{event_df:}{all information regarding events:
\describe{
\item{event:}{}
\item{FC_HL:}{}
\item{FC_intensity:}{}
\item{FC_HL_FC_intensity:}{}
\item{p_adjusted:}{}
\item{velocity_ratio:}{}
\item{p_value:}{}
\item{feature_type:}{}
\item{gene:}{}
\item{locus_tag:}{}
\item{strand:}{The bin/probe specific strand}
\item{TU:}{The overarching transcription unit}
\item{segment_1:}{}
\item{segment_2:}{}
\item{event_position:}{}
\item{event_duration:}{}
\item{gap_fragments:}{}
\item{features:}{}
}
}
\item{TI_df:}{all information regarding TI are depicted below:
\describe{
\item{event:}{}
\item{TI_fragment:}{}
\item{TI_termination_factor:}{}
\item{p_value:}{}
\item{p_adjusted:}{}
\item{feature_type:}{}
\item{gene:}{}
\item{locus_tag:}{}
\item{strand:}{The bin/probe specific strand}
\item{TU:}{The overarching transcription unit}
\item{features:}{}
\item{event_position:}{}
\item{position_1:}{}
\item{position_2:}{}
}
}
}
}
\source{
\url{https://github.com/CyanolabFreiburg/rifi}
}
\usage{
data(summary_e_coli)
}
\description{
The result of rifi_summary for E.coli example data
A SummarizedExperiment containing the output of rifi_stats as an
extention to rowRanges
}
\keyword{datasets}
