% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_pen_obj.r
\name{viz_pen_obj}
\alias{viz_pen_obj}
\title{viz_pen_obj provides an optional visualization of any penalty object
created by make_pen. the function aan be customized to show only
the n = top_i top results.}
\usage{
viz_pen_obj(obj, top_i = nrow(obj[[3]][[1]]) * ncol(obj[[3]][[1]]))
}
\arguments{
\item{obj}{object: penalty object(make_pen output)}

\item{top_i}{integer: the number of top results visualized. Default is all.}
}
\value{
A visualization of the penalty object
}
\description{
viz_pen_obj provides an optional visualization of any penalty object
created by make_pen. the function aan be customized to show only
the n = top_i top results.
}
\examples{
data(penalties_e_coli)
viz_pen_obj(penalties_e_coli$pen_obj_delay,25)

}
