% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{wrapper_e_coli}
\alias{wrapper_e_coli}
\title{A list corresponding to a gff file for E.coli example data
A list containing all necessary information from a gff file for rifi_summary
and visualization. The list is stored as metadata in rifi_stats output.}
\format{
A list with 2 items:
\describe{
\item{data annotation:}{a data frame with 4452 rows and 6 variables
\describe{
\item{region:}{the region from the gff file}
\item{start:}{the start of the annotation}
\item{end:}{the end of the annotation}
\item{strand:}{the strand of the annotation}
\item{gene:}{the annotated gene name}
\item{locus_tag:}{the annotated locus tag}
}
}
\item{genome length:}{a numeric vector containing the length of the genome}
}

The result of rifi_wrapper for E.coli example data
A list of SummarizedExperiment containing the output of rifi_wrapper. The
list contains 6 elements of SummarizedExperiment output of rifi_preprocess,
rifi_fit, rifi_penalties, rifi_fragmentation, rifi_stats and rifi_summary.
The plot is generated from rifi_visualization. for more detail, please refer
to each function separately.#'
}
\source{
\url{https://github.com/CyanolabFreiburg/rifi}
}
\usage{
data(wrapper_e_coli)
}
\description{
A list corresponding to a gff file for E.coli example data
A list containing all necessary information from a gff file for rifi_summary
and visualization. The list is stored as metadata in rifi_stats output.
}
\keyword{datasets}
