% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{wrapper_summary_synechocystis_6803}
\alias{wrapper_summary_synechocystis_6803}
\title{The result of rifi_wrapper for summary_synechocystis_6803 example data
A list of SummarizedExperiment containing the output of rifi_wrapper. The
list contains 6 elements of SummarizedExperiment output of rifi_preprocess,
rifi_fit, rifi_penalties, rifi_fragmentation, rifi_stats and rifi_summary.
The plot is generated from rifi_visualization. for more detail, please refer
to each function separately.}
\format{
An object of class \code{list} of length 6.
}
\source{
\url{https://github.com/CyanolabFreiburg/rifi}
}
\usage{
data(wrapper_summary_synechocystis_6803)
}
\description{
The result of rifi_wrapper for summary_synechocystis_6803 example data
A list of SummarizedExperiment containing the output of rifi_wrapper. The
list contains 6 elements of SummarizedExperiment output of rifi_preprocess,
rifi_fit, rifi_penalties, rifi_fragmentation, rifi_stats and rifi_summary.
The plot is generated from rifi_visualization. for more detail, please refer
to each function separately.
}
\keyword{datasets}
