% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_ancova.r
\name{apply_ancova}
\alias{apply_ancova}
\title{apply_ancova: is a statistical test to check variances between 2 segments
showing pausing site (ps) or internal starting site (ITSS) independently.
apply_ancova: is a statistical test to check if fragments showing ps and
ITSS events have significant slope using Ancova test.
The function uses ancova test. Ancova is applied when the data contains
independent variables, dependent variables and covariant variables.
In this case, segments are independent variables, position is the dependent
variable and the delay is the covariant.
The dataframe is prepared as depicted below. The lm fit is applied and
p_value is extracted
delay  position  segment
S1
S1
S1
S1
S2
S2
S2
S2}
\usage{
apply_ancova(data)
}
\arguments{
\item{data}{dataframe: the probe based data frame.}
}
\value{
the probe data frame with the columns regarding statistics:
\describe{
\item{ID:}{The bin/probe specific ID}
\item{position:}{The bin/probe specific position}
\item{strand:}{The bin/probe specific strand}
\item{delay:}{The delay value of the bin/probe}
\item{intercept:}{The vintercept of fit through the respective delay
fragment}
\item{slope:}{The slope of the fit through the respective delay
fragment}
\item{pausing_site:}{}
\item{iTSS_I:}{}
\item{ps_ts_fragment:}{}
\item{event_ps_itss_p_value_Ttest:}{}
\item{p_value_slope:}{}
\item{delay_frg_slope:}{}
\item{velocity_ratio:}{}
}
}
\description{
apply_ancova: is a statistical test to check variances between 2 segments
showing pausing site (ps) or internal starting site (ITSS) independently.
apply_ancova: is a statistical test to check if fragments showing ps and
ITSS events have significant slope using Ancova test.
The function uses ancova test. Ancova is applied when the data contains
independent variables, dependent variables and covariant variables.
In this case, segments are independent variables, position is the dependent
variable and the delay is the covariant.
The dataframe is prepared as depicted below. The lm fit is applied and
p_value is extracted
delay  position  segment
S1
S1
S1
S1
S2
S2
S2
S2
}
\examples{
data(stats_minimal)
apply_ancova(data = stats_minimal)

}
