% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_event_position.r
\name{apply_event_position}
\alias{apply_event_position}
\title{apply_event_position is a short version of apply_Ttest_delay function
to extract event time duration for pausing site or iTSS.}
\usage{
apply_event_position(inp)
}
\arguments{
\item{inp}{SummarizedExperiment: the input data frame with correct format.}
}
\value{
the SummarizedExperiment with the columns regarding statistics:
\describe{
\item{event_position:}{Integer, position of the event added to the input}
}
}
\description{
apply_event_position adds a new column with the duration.
}
\examples{
data(stats_minimal)
apply_event_position(inp = stats_minimal)

}
