% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_event_position.r
\name{apply_event_position}
\alias{apply_event_position}
\title{apply_event_position: is a short version of apply_Ttest_delay function
to extract event time duration as pausing site or iTSS happens.}
\usage{
apply_event_position(inp)
}
\arguments{
\item{inp}{SummarizedExperiment: the input data frame with correct format.}
}
\value{
the probe data frame with the columns regarding statistics:
\describe{
\item{ID:}{The bin/probe specific ID}
\item{position:}{The bin/probe specific position}
\item{strand:}{The bin/probe specific strand}
\item{delay:}{The delay value of the bin/probe}
\item{pausing_site:}{}
\item{iTSS_I:}{}
\item{ps_ts_fragment:}{}
\item{event_ps_itss_p_value_Ttest:}{}
\item{p_value_slope:}{}
\item{delay_frg_slope:}{}
\item{velocity_ratio:}{}
\item{event_position:}{}
}
}
\description{
apply_event_position adds a new column with the duration.
}
\examples{
data(stats_minimal)
apply_event_position(inp = stats_minimal)

}
