% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_t_test.r
\name{apply_t_test}
\alias{apply_t_test}
\title{of half-life (HL) fragments and the fold-change
intensity fragments respectively are significant.}
\usage{
apply_t_test(inp, threshold = 300)
}
\arguments{
\item{inp}{SummarizedExperiment: the input data frame with correct format.}

\item{threshold}{integer: threshold.}
}
\value{
the SummarizedExperiment with the columns regarding statistics:
\describe{
\item{FC_HL:}{Integer, the fold change value of 2 HL fragments}
\item{FC_fragment_HL:}{String, the fragments corresponding to HL fold change}
\item{p_value_HL:}{Integer, the p_value added to the input of 2 HL fragments}
\item{FC_intensity:}{Integer, the fold change value of 2 intensity fragments}
\item{FC_fragment_intensity:}{String, the fragments corresponding to
intensity fold change}
\item{p_value_intensity:}{Integer, the p_value added to the input of 2
intensity fragments}
}
}
\description{
apply_t_test compares the mean of two neighboring fragments within the
same TU to check if the fold-change is significant.
Fragments with distance above threshold are not subjected to t-test.
Dataframes with less than 3 rows are excluded.
}
\details{
The functions used are:
\enumerate{
\item fragment_function: checks number of fragments inside TU, less
than 2 are excluded otherwise they are gathered for analysis.
\item t_test_function: excludes dataframes with less than 3 rows,
makes fold-change and apply t-test, assign fragments names
and ratio, add columns with the corresponding p_values.
}
}
\examples{
data(stats_minimal)
apply_t_test(inp = stats_minimal, threshold = 300)

}
