% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_t_test_ti.r
\name{apply_t_test_ti}
\alias{apply_t_test_ti}
\title{within the same TU.}
\usage{
apply_t_test_ti(inp)
}
\arguments{
\item{inp}{SummarizedExperiment: the input data frame with correct format.}
}
\value{
the SummarizedExperiment with the columns regarding statistics:
\describe{
\item{p_value_TI:}{Integer, the p_value added to the input}
\item{TI_fragments_p_value:}{String, the fragments subjected to statistical test}
}
}
\description{
apply_t_test_ti uses the statistical t_test to check if two neighboring
TI fragments are significant.
}
\examples{
data(stats_minimal)
apply_t_test_ti(inp = stats_minimal)

}
