% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_Ttest_delay.r
\name{apply_Ttest_delay}
\alias{apply_Ttest_delay}
\title{segments showing pausing site (ps) and internal
starting site (ITSS) independently}
\usage{
apply_Ttest_delay(inp)
}
\arguments{
\item{inp}{SummarizedExperiment: the input data frame with correct format.}
}
\value{
the SummarizedExperiment with the columns regarding statistics:
\describe{
\item{event_ps_itss_p_value_Ttest:}{Integer, the p_value added to the input}
}
}
\description{
apply_Ttest_delay: is a statistical test to check the significance
of the point between 2 segments showing pausing site (ps) and
internal starting site (ITSS) independently.
apply_Ttest_delay uses t-test. The last point from the first segment
and the first point from the second segment are selected and added
to the residuals of each model. The sum is subjected to t-test.
}
\examples{
data(stats_minimal)
apply_Ttest_delay(inp = stats_minimal)

}
