% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.r
\name{check_input}
\alias{check_input}
\title{check_input: reviews the input given by the user.
'check_input' stops the operation if the input data frame has severe faults.
Less severe faults lead to the removal of wrong IDs and a warnings describing
the problem.
The input data frame with all possible replicates need to have the following
columns:
"0","'time point x'",...,"'time point n'","ID","position","strand"!
It is required that the names referring to the time points can be converted
to numeric.
IDs can be given as character (except ",","|","_"), or numeric, but need to
refer to a unique position/strand combination.
Strand information need to be given by "+" and "-".
The columns "ID","position","strand" must be the last three columns in this
order!
The relative intensity for all time points should be numeric.
The relative intensity for the first time point cannot be 0 or NA.}
\usage{
check_input(inp, thrsh = 0)
}
\arguments{
\item{inp}{data frame: the input data frame with correct format.}

\item{thrsh}{numeric: the minimal allowed intensity for time point "0"...
...advised to be kept at 0! Default is 0.}
}
\value{
the input data frame, with the filtration column added:
\describe{
\item{...}{all timepoints}
\item{ID:}{unique IDs}
\item{position:}{genome positions}
\item{strand:}{strand information}
\item{filtration:}{indicator wether the replicate is filtered or not}
}
}
\description{
check_input: reviews the input given by the user.
'check_input' stops the operation if the input data frame has severe faults.
Less severe faults lead to the removal of wrong IDs and a warnings describing
the problem.
The input data frame with all possible replicates need to have the following
columns:
"0","'time point x'",...,"'time point n'","ID","position","strand"!
It is required that the names referring to the time points can be converted
to numeric.
IDs can be given as character (except ",","|","_"), or numeric, but need to
refer to a unique position/strand combination.
Strand information need to be given by "+" and "-".
The columns "ID","position","strand" must be the last three columns in this
order!
The relative intensity for all time points should be numeric.
The relative intensity for the first time point cannot be 0 or NA.
}
\examples{
data(example_input_minimal)
check_input(inp = example_input_minimal, thrsh = 0)
}
