% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_summary.r
\name{dataframe_summary}
\alias{dataframe_summary}
\title{dataframe_summary: creates two tables relating gene annotation to fragments.
dataframe_summary creates two tables summary of segments and their
half-lives. The first output is bin/probe features and the second one is
intensity fragment based.
The dataframe_summary creates one table with feature_type, gene, locus_tag,
position, strand, TU, delay_fragment, HL_fragment,
half_life, intensity_fragment, intensity and velocity. The second table is
similar to the first one but in compact form.
It contains the same columns, the only difference is on position where a
start and end position are indicated separately.
Strand is indicated in case of stranded data to select the corresponding
positions.}
\usage{
dataframe_summary(data, input)
}
\arguments{
\item{data}{SummarizedExperiment: the input data frame with correct format.}

\item{input}{dataframe: dataframe from event_dataframe function.}
}
\value{
\item{bin_df:}{all information regarding bins:
\describe{
\item{feature_type:}{}
\item{gene:}{}
\item{locus_tag:}{}
\item{TU:}{The overarching transcription unit}
\item{delay_fragment:}{The delay fragment the bin belongs to}
\item{HL_fragment:}{The half-life fragment the bin belongs to}
\item{half_life:}{The half-life of the bin/probe}
\item{intensity_fragment:}{The intensity fragment the bin belongs to}
\item{intensity:}{The relative intensity at time point 0}
\item{velocity:}{The velocity value of the bin}
}
}
\item{frag_df:}{all information regarding fragments:
\describe{
\item{feature_type:}{}
\item{gene:}{}
\item{locus_tag:}{}
\item{strand:}{The bin/probe specific strand}
\item{TU:}{The overarching transcription unit}
\item{delay_fragment:}{The delay fragment the bin belongs to}
\item{HL_fragment:}{The half-life fragment the bin belongs to}
\item{half_life:}{The half-life of the fragment}
\item{intensity_fragment:}{The intensity fragment the bin belongs to}
\item{intensity:}{The relative intensity at time point 0}
\item{velocity:}{The velocity value of the respective delay fragment}
}
}
}
\description{
dataframe_summary: creates two tables relating gene annotation to fragments.
dataframe_summary creates two tables summary of segments and their
half-lives. The first output is bin/probe features and the second one is
intensity fragment based.
The dataframe_summary creates one table with feature_type, gene, locus_tag,
position, strand, TU, delay_fragment, HL_fragment,
half_life, intensity_fragment, intensity and velocity. The second table is
similar to the first one but in compact form.
It contains the same columns, the only difference is on position where a
start and end position are indicated separately.
Strand is indicated in case of stranded data to select the corresponding
positions.
}
\examples{
data(stats_minimal)
data(res_minimal)
dataframe_summary(data = stats_minimal, input = res_minimal)
}
