% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_summary.r
\name{dataframe_summary}
\alias{dataframe_summary}
\title{dataframe_summary: creates two tables relating gene annotation to fragments.
dataframe_summary creates two tables summary of segments and their
half-lives. The first output is bin/probe features and the second one is
intensity fragment based.
The dataframe_summary creates one table with feature_type, gene, locus_tag,
position, strand, TU, delay_fragment, HL_fragment,
half_life, intensity_fragment, intensity and velocity. The second table is
similar to the first one but in compact form.
It contains the same columns, the only difference is on position where a
start and end position are indicated separately.
Strand is indicated in case of stranded data to select the corresponding
positions.}
\usage{
dataframe_summary(inp, input)
}
\arguments{
\item{inp}{SummarizedExperiment: the input data frame with correct format.}

\item{input}{dataframe: the probe based data frame with events and
gene annotation.}
}
\value{
WIP
}
\description{
dataframe_summary: creates two tables relating gene annotation to fragments.
dataframe_summary creates two tables summary of segments and their
half-lives. The first output is bin/probe features and the second one is
intensity fragment based.
The dataframe_summary creates one table with feature_type, gene, locus_tag,
position, strand, TU, delay_fragment, HL_fragment,
half_life, intensity_fragment, intensity and velocity. The second table is
similar to the first one but in compact form.
It contains the same columns, the only difference is on position where a
start and end position are indicated separately.
Strand is indicated in case of stranded data to select the corresponding
positions.
}
\examples{
data(stats_minimal)
data(annot_g_minimal)
input <- event_dataframe(data = as.data.frame(rowRanges(stats_minimal)),
data_annotation = annot_g_minimal[[1]])
dataframe_summary(inp = stats_minimal, input = input)
}
