% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_summary_events_velocity.r
\name{dataframe_summary_events_velocity}
\alias{dataframe_summary_events_velocity}
\title{dataframe_summary_events_velocity creates one table with all events between
the segments.
The dataframe_summary_events_velocity creates one table with the following
columns: event, features, p_value, event_position, event_duration,
position, region, gene, locus_tag, strand, TU, segment_1, segment_2, length,
velocity_ratio.
The columns are:
\enumerate{
\item event: event type, pausing site, iTSS_I, iTSS_II, Termination, HL_event,
Int_event, HL_Int_event and velocity_change.
\item velocity_ratio: ratio between any two fragment where the event happen.
\item p_value: depending on the event, t-test, manova test p_value is assigned.
\item feature_type: indicated on the output data frame as region, are the
feature type covering the event.
\item gene: gene covering the event.
\item locus_tag: locus_tag covering the event.
\item strand: +/- indicated in case of stranded data.
\item TU: TU covering the event.
\item segment_1: the first segment of the event, includes the segment, TU,
delay fragment in case of ps or iTSS_I. The rest of the events include HL
fragment and could be extended
intensity fragment.
\item segment_2: same description as segment_1 but is the second fragment of
the event.
\item event_position: the position of event, calculated dividing the last
position of the first fragment and the first position of the next fragment
on 2.
\item event_duration: the difference (min) between 2 delay fragment when ps
or iTSS_I happen.
\item gap_fragments: length in position (nt), calculated by the difference
between the last position of the first fragment and the first position of
the second fragment.
\item features: number of segment involved on the event.
}}
\usage{
dataframe_summary_events_velocity(data, data_annotation)
}
\arguments{
\item{data}{SummarizedExperiment: the input data frame with correct format.}

\item{data_annotation}{dataframe: dataframe from processed gff3 file.}
}
\value{
\describe{
\item{event:}{}
\item{p_value:}{}
\item{p_adjusted:}{}
\item{event_position:}{}
\item{velocity_ratio:}{}
\item{feature_type:}{}
\item{gene:}{}
\item{locus_tag:}{}
\item{strand:}{The bin/probe specific strand}
\item{TU:}{The overarching transcription unit}
\item{segment_1:}{}
\item{segment_2:}{}
\item{event_duration:}{}
\item{gap_fragments:}{}
\item{features:}{}
}
}
\description{
dataframe_summary_events_velocity creates one table with all events between
the segments.
The dataframe_summary_events_velocity creates one table with the following
columns: event, features, p_value, event_position, event_duration,
position, region, gene, locus_tag, strand, TU, segment_1, segment_2, length,
velocity_ratio.
The columns are:
\enumerate{
\item event: event type, pausing site, iTSS_I, iTSS_II, Termination, HL_event,
Int_event, HL_Int_event and velocity_change.
\item velocity_ratio: ratio between any two fragment where the event happen.
\item p_value: depending on the event, t-test, manova test p_value is assigned.
\item feature_type: indicated on the output data frame as region, are the
feature type covering the event.
\item gene: gene covering the event.
\item locus_tag: locus_tag covering the event.
\item strand: +/- indicated in case of stranded data.
\item TU: TU covering the event.
\item segment_1: the first segment of the event, includes the segment, TU,
delay fragment in case of ps or iTSS_I. The rest of the events include HL
fragment and could be extended
intensity fragment.
\item segment_2: same description as segment_1 but is the second fragment of
the event.
\item event_position: the position of event, calculated dividing the last
position of the first fragment and the first position of the next fragment
on 2.
\item event_duration: the difference (min) between 2 delay fragment when ps
or iTSS_I happen.
\item gap_fragments: length in position (nt), calculated by the difference
between the last position of the first fragment and the first position of
the second fragment.
\item features: number of segment involved on the event.
}
}
\examples{
if(!require(SummarizedExperiment)){
suppressPackageStartupMessages(library(SummarizedExperiment))
}
data(stats_minimal)
dataframe_summary_events_velocity(data = stats_minimal,
data_annotation = metadata(stats_minimal)$annot[[1]])

}
