% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_summary_TI.r
\name{dataframe_summary_TI}
\alias{dataframe_summary_TI}
\title{dataframe_summary_TI creates one table with all TI fragments, p_value and
the coordinates.
The dataframe_summary creates one table with the following columns: event,
TI_fragment, TI_factor, TI_fragments_TU, p_value, feature_type,
gene, locus_tag, strand, TU, features, event_position, position_1 and
position_2.
The columns are:
\enumerate{
\item event: event type, transcription interference.
\item TI_fragment: Transcription interference fragment.
\item TI_factor: Transcription interference factor.
\item TI_fragments_TU: Transcription interference fragments included on the
TU.
\item p_value: TI p_value between two successive fragments is assigned.
\item feature_type: indicated on the output data frame as region, are the
feature type covering the TI.
\item gene: the genes covering the TI.
\item locus_tag: the locus_tags covering the TI.
\item strand: +/- indicated in case of stranded data.
\item TU: TU covering the TI.
\item features: number of segment TI involved on a TU.
\item event_position : position between two TI fragments.
\item position_1 : the first position of TI fragment, if 2 fragments, first
position is from the first fragment.
\item position_2 : the last position of TI fragment, if 2 fragments, last
position is from the second fragment.
}}
\usage{
dataframe_summary_TI(data, input)
}
\arguments{
\item{data}{SummarizedExperiment: the input data frame with correct format.}

\item{input}{dataframe: dataframe from event_dataframe function.}
}
\value{
WIP
}
\description{
dataframe_summary_TI creates one table with all TI fragments, p_value and
the coordinates.
The dataframe_summary creates one table with the following columns: event,
TI_fragment, TI_factor, TI_fragments_TU, p_value, feature_type,
gene, locus_tag, strand, TU, features, event_position, position_1 and
position_2.
The columns are:
\enumerate{
\item event: event type, transcription interference.
\item TI_fragment: Transcription interference fragment.
\item TI_factor: Transcription interference factor.
\item TI_fragments_TU: Transcription interference fragments included on the
TU.
\item p_value: TI p_value between two successive fragments is assigned.
\item feature_type: indicated on the output data frame as region, are the
feature type covering the TI.
\item gene: the genes covering the TI.
\item locus_tag: the locus_tags covering the TI.
\item strand: +/- indicated in case of stranded data.
\item TU: TU covering the TI.
\item features: number of segment TI involved on a TU.
\item event_position : position between two TI fragments.
\item position_1 : the first position of TI fragment, if 2 fragments, first
position is from the first fragment.
\item position_2 : the last position of TI fragment, if 2 fragments, last
position is from the second fragment.
}
}
\examples{
data(stats_minimal)
data(res_minimal)
dataframe_summary_TI(data = stats_minimal, input = res_minimal)

}
