% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_dataframe.r
\name{event_dataframe}
\alias{event_dataframe}
\title{event_dataframe creates a dataframe only with events for segments and genes.
The function used are:
position_function: adds the specific position of ps or iTSS event.
annotation_function_event: adds the events to the annotated genes.
gff3 file has to be supplied. Strand is indicated in case of stranded data
The event_dataframe selects columns with statistical features. ID, position,
strand and TU columns are required.
Two major dataframe are generated, df gathers t-test and Manova test and df1
gathers ps and ITSS with the corresponding features. df selects only unique
intensity fragments since they are the lowest on the hierarchy.
One new column is added to df, "synthesis_ratio_event", it corresponds
to the FC-HL/FC-intensity and assignment of an event to the synthesis ratio
respectively. df adds a new column to indicate the position of the ps or
ITSS event.}
\usage{
event_dataframe(data, data_annotation)
}
\arguments{
\item{data}{dataframe: the probe based data frame.}

\item{data_annotation}{dataframe: the coordinates are extracted from gff3}
}
\value{
WIP
}
\description{
event_dataframe creates a dataframe only with events for segments and genes.
The function used are:
position_function: adds the specific position of ps or iTSS event.
annotation_function_event: adds the events to the annotated genes.
gff3 file has to be supplied. Strand is indicated in case of stranded data
The event_dataframe selects columns with statistical features. ID, position,
strand and TU columns are required.
Two major dataframe are generated, df gathers t-test and Manova test and df1
gathers ps and ITSS with the corresponding features. df selects only unique
intensity fragments since they are the lowest on the hierarchy.
One new column is added to df, "synthesis_ratio_event", it corresponds
to the FC-HL/FC-intensity and assignment of an event to the synthesis ratio
respectively. df adds a new column to indicate the position of the ps or
ITSS event.
}
\examples{
if(!require(SummarizedExperiment)){
suppressPackageStartupMessages(library(SummarizedExperiment))
}
data(stats_minimal)
event_dataframe(data = stats_minimal,
data_annotation = metadata(stats_minimal)$annot[[1]])

}
