% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{example_input_minimal}
\alias{example_input_minimal}
\title{An artificial example input data frame.
A data set containing information about the intensities at all time points,
as well as position, strand and ID information.}
\format{
An object of class \code{RangedSummarizedExperiment} with 4 rows and 33 columns.
}
\source{
\url{https://github.com/CyanolabFreiburg/Transcriptome_wide_decay}
}
\usage{
data(example_input_minimal)
}
\description{
An artificial example input data frame.
A data set containing information about the intensities at all time points,
as well as position, strand and ID information.
}
\keyword{datasets}
