% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{fit_e_coli}
\alias{fit_e_coli}
\title{The result of rifi_fit for E.coli example data
A summarizedExperiment containing the output from rifi_fit as an extension of
rowRanges and metadata.}
\format{
Three data frames with 290 rows and 10 variables, 155 rows
and 5 variables, and 135 rows and 9 variables are generated. The columns of
the first data frame are added to the rowRanges and the rest are added as
metadata.
\describe{
\item{inp:}{The summarizedExperiment:
\describe{
\item{ID:}{The bin/probe specific ID}
\item{position:}{The bin/probe specific position}
\item{intensity:}{The relative intensity at time point 0}
\item{probe_TI:}{An internal value to determine which fitting model
is applied}
\item{flag:}{Information on which fitting model is applied}
\item{postion_segment:}{The position based segment}
\item{delay:}{The delay value of the bin/probe}
\item{half_life:}{The half-life of the bin/probe}
\item{TI_termination_factor:}{The termination factor of the bin/probe}
}
}
\item{fit_obj_STD:}{the fit object for the standard fit:
\describe{
\item{ID:}{The bin/probe specific ID}
\item{delay:}{The delay value of the bin/probe}
\item{half_life:}{The half-life of the bin/probe}
\item{inty_S0:}{The relative intensity at time point 0}
\item{intyf:}{The background value of the fit}
}
}
\item{fit_obj_TI:}{the fit object for the TI fit:
\describe{
\item{delay:}{The delay value of the bin/probe}
\item{ti_delay:}{The ti-delay value of the bin/probe}
\item{half_life:}{The half-life of the bin/probe}
\item{ti_value:}{The ti-value of the bin/probe}
\item{TI_termination_factor:}{The termination factor of the bin/probe}
\item{synthesis_rate:}{The synthesis rate of the bin/probe}
\item{TI_background:}{The background value of the fit}
\item{position:}{The bin/probe specific position}
\item{ID:}{The bin/probe specific ID}
}
}
}
}
\source{
\url{https://github.com/CyanolabFreiburg/rifi}
}
\usage{
data(fit_e_coli)
}
\description{
The result of rifi_fit for E.coli example data
A summarizedExperiment containing the output from rifi_fit as an extension of
rowRanges and metadata.
}
\keyword{datasets}
