% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{fit_minimal}
\alias{fit_minimal}
\title{The artificial result of rifi_fit for artificial example data
A data frame containing the output from rifi_fit.}
\format{
An object of class \code{RangedSummarizedExperiment} with 4 rows and 33 columns.
}
\source{
\url{https://github.com/CyanolabFreiburg/Transcriptome_wide_decay}
}
\usage{
data(fit_minimal)
}
\description{
The artificial result of rifi_fit for artificial example data
A data frame containing the output from rifi_fit.
}
\keyword{datasets}
