% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_change.r
\name{fold_change}
\alias{fold_change}
\title{fold_change: it sets a fold-change ratio between the neighboring fragments
of Half-life (HL) and intensity.
fold_change sets fold change on intensity and fold change HL fragments of
two successive fragments. Two intensity fragments could belong to one HL
fragment.
This function sets first the borders using the position and applies the fold
change ratio between the neighboring fragments of HL and those from intensity
(intensity frgA/intensity frgB/half-life frgA/half-life frgB). All grepped
fragments are from the same TU excluding outliers.}
\usage{
fold_change(inp)
}
\arguments{
\item{inp}{SummarizedExperiment: the input data frame with correct format.}
}
\value{
the probe data frame with the columns regarding statistics:
\describe{
\item{ID:}{The bin/probe specific ID}
\item{position:}{The bin/probe specific position}
\item{strand:}{The bin/probe specific strand}
\item{intensity:}{The relative intensity at time point 0}
\item{half_life:}{The half-life of the bin/probe}
\item{HL_fragment:}{The half-life fragment the bin belongs to}
\item{HL_mean_fragment:}{The mean half-life value of the respective
half-life fragment}
\item{intensity_fragment:}{The intensity fragment the bin belongs to}
\item{intensity_mean_fragment:}{The mean intensity value of the respective
intensity fragment}
\item{TU:}{The overarching transcription unit}
\item{pausing_site:}{}
\item{iTSS_I:}{}
\item{ps_ts_fragment:}{}
\item{event_ps_itss_p_value_Ttest:}{}
\item{p_value_slope:}{}
\item{delay_frg_slope:}{}
\item{velocity_ratio:}{}
\item{event_duration:}{}
\item{event_position:}{}
\item{FC_HL:}{}
\item{FC_fragment_HL:}{}
\item{p_value_HL:}{}
\item{FC_intensity:}{}
\item{FC_fragment_intensity:}{}
\item{p_value_intensity:}{}
\item{FC_HL_intensity:}{}
\item{FC_HL_intensity_fragment:}{}
\item{FC_HL_adapted:}{}
}
}
\description{
The function used is:
synthesis_r_Function: assigns events depending on the ratio between HL and
intensity of two consecutive fragments.
intensity(int) = synthesis rate(k)/decay(deg) (steady state),
int1/int2 = k1/deg1*deg2/k2
int1 * (deg1/int2) * deg2 = k1/k2 => synthesis ratio. In case of synthesis
ratio is:
synthesis ratio > 1 -> New start
synthesis ratio < 1 -> Termination
}
\examples{
data(stats_minimal)
fold_change(inp = stats_minimal)

}
