% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_change.r
\name{fold_change}
\alias{fold_change}
\title{fold_change sets fold change on intensity and fold change HL fragments of
two successive fragments. Two intensity fragments could belong to one HL
fragment.
This function sets first the borders using the position and applies the fold
change ratio between the neighboring fragments of HL and those from intensity
log2(intensity frgA/intensity frgB/half-life frgA/half-life frgB). All grepped
fragments are from the same TU excluding outliers.}
\usage{
fold_change(inp)
}
\arguments{
\item{inp}{SummarizedExperiment: the input data frame with correct format.}
}
\value{
the SummarizedExperiment with the columns regarding statistics:
\describe{
\item{synthesis_ratio:}{Integer, the value correspomding to synthesis rate}
\item{synthesis_ratio_event:}{String, the event assigned by synthesis rate either
Termination or iTSS}
\item{FC_HL_intensity:}{Integer, the value corresponding to HL and intensity fold change}
\item{FC_HL_intensity_fragment:}{String, the fragments corresponding to intensity
and HL fold change}
\item{FC_HL_adapted:}{Integer, the fold change of half-life/ fold change of intensity,
position of the half-life fragment is adapted to intensity fragment}
}
}
\description{
The function used is:
synthesis_r_Function: assigns events depending on the ratio between HL and
intensity of two consecutive fragments.
intensity(int) = synthesis rate(k)/decay(deg) (steady state),
int1/int2 = k1/deg1*deg2/k2
int1 * (deg1/int2) * deg2 = k1/k2 => synthesis ratio.
In case of synthesis ratio is:
synthesis ratio > 0 -> New start
synthesis ratio < 0 -> Termination
}
\examples{
data(stats_minimal)
fold_change(inp = stats_minimal)

}
