% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragment_TI.r
\name{fragment_TI}
\alias{fragment_TI}
\title{\subsection{=========================================================================
fragment_TI}{

fragment_TI performs the TI fragmentation
}}
\usage{
fragment_TI(inp, cores = 1, pen, pen_out)
}
\arguments{
\item{inp}{SummarizedExperiment: the input data frame with correct format.}

\item{cores}{cores: integer: the number of assigned cores for the task.}

\item{pen}{numeric: an internal parameter for the dynamic programming.
Higher values result in fewer fragments. Default is the auto generated value.}

\item{pen_out}{numeric: an internal parameter for the dynamic programming.
Higher values result in fewer allowed outliers. Default is the auto generated
value.}
}
\value{
the SummarizedExperiment object: with TI_termination_fragment and
TI_termination_mean_fragment added to the rowRanges.
}
\description{
fragment_TI makes TI_fragments based on TUs and assigns all gathered
information to the SummarizedExperiment object. The columns
TI_termination_fragment" and the TI_mean_termination_factor are added.
}
\details{
The function used is:

.score_fun_ave.

The input is the SummarizedExperiment object.

pen is the penalty for new fragments in the dynamic programming, pen_out is
the outlier penalty.
}
\examples{
data(fragmentation_minimal)
fragment_TI(inp = fragmentation_minimal, cores = 2, pen = 2, pen_out = 1)

}
