% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{penalties_minimal}
\alias{penalties_minimal}
\title{The result of rifi_penalties for artificial example data
A vector containing the output from rifi_penalties, the logbook.}
\format{
A numeric vector of length 8:
\describe{
\item{logbook:}{The logbook vector containing all penalty information}
}
}
\source{
\url{https://github.com/CyanolabFreiburg/Transcriptome_wide_decay}
}
\usage{
data(penalties_minimal)
}
\description{
The result of rifi_penalties for artificial example data
A vector containing the output from rifi_penalties, the logbook.
}
\keyword{datasets}
