% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ps_itss.r
\name{predict_ps_itss}
\alias{predict_ps_itss}
\title{(ITSS) between delay fragments.}
\usage{
predict_ps_itss(inp, maxDis = 300)
}
\arguments{
\item{inp}{SummarizedExperiment: the input data frame with correct format.}

\item{maxDis}{integer: the maximal distance allowed between two successive
fragments.}
}
\value{
the SummarizedExperiment with the columns regarding statistics:
\describe{
\item{pausing_site:}{Boolean, presence or absence of pausing_site event (ps)}
\item{iTSS_I:}{Boolean, presence or absence of internal starting site event (iTSS_I)}
\item{ps_ts_fragment:}{String, fragments involved on the event}
\item{event_duration:}{Integer, the duration between two delay fragments}
}
}
\description{
predict_ps_itss predicts ps and ITSS within the same TU. Neighboring delay
segments are compared to each other by positioning the intercept of the
second segment into the first segment using slope and intercept coefficients.
predict_ps_itss uses 3 steps to identify ps and ITSS:
\enumerate{
\item select unique TU.
\item select from the input dataframe the columns: ID, position, strand, delay.
delay fragment, TU and slope coordinates, velocity_fragment and intercept.
\item select delay segments in the TU.
\item loop into all delay segments and estimate the coordinates of the last
point of the first segment using the coefficients of the second segment
and vice versa. We get two predicted positions, the difference between
them is compared to the threshold.
In case the strand is "-", additional steps are added:
The positions of both segments are ordered from the last position to the
first one.
All positions are merged in one column and subtracted from the maximum
position. the column is split in 2. The first and second correspond to
the positions of the first and second segments respectively.
Both segments are subjected to lm fit and the positions predicted are used
on the same way as the opposite strand.
If the difference between the positions predicted is lower than negative
threshold, ps is assigned otherwise, and if the difference is higher than
the positive threshold, ITSS is assigned.
}
}
\examples{
data(fragmentation_minimal)
predict_ps_itss(inp = fragmentation_minimal, maxDis = 300) 
}
