% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{preprocess_e_coli}
\alias{preprocess_e_coli}
\title{The result of rifi_preprocess for E.coli example data
A SummarizedExperiment containing the output from rifi_penalties including
the logbook and the four penalty objects as metadata.
A list containing the output from rifi_preprocess, including the inp
and the modified input_df.}
\format{
A SummarizedExperiment:
\describe{
\item{inp:}{the SummarizedExperiment:
\describe{
\item{ID:}{The bin/probe specific ID}
\item{position:}{The bin/probe specific position}
\item{intensity:}{The relative intensity at time point 0}
\item{probe_TI:}{An internal value to determine which fitting model is
applied}
\item{flag:}{Information on which fitting model is applied}
\item{postion_segment:}{The position based segment}
}
}
\item{fit_obj_TI:}{the fit object for the TI fit:
\describe{
\item{0:}{relative intensities at 0 min}
\item{1:}{relative intensities at 1 min}
\item{10:}{relative intensities at 10 min}
\item{15:}{relative intensities at 15 min}
\item{2:}{relative intensities at 2 min}
\item{20:}{relative intensities at 20 min}
\item{3:}{relative intensities at 3 min}
\item{4:}{relative intensities at 4 min}
\item{5:}{relative intensities at 5 min}
\item{6:}{relative intensities at 6 min}
\item{8:}{relative intensities at 8 min}
\item{ID:}{unique IDs}
\item{position:}{genome positions}
\item{filtration:}{indicator wether the replicate is filtered or not}
}
}
}
}
\source{
\url{https://github.com/CyanolabFreiburg/rifi}
}
\usage{
data(preprocess_e_coli)
}
\description{
The result of rifi_preprocess for E.coli example data
A SummarizedExperiment containing the output from rifi_penalties including
the logbook and the four penalty objects as metadata.
A list containing the output from rifi_preprocess, including the inp
and the modified input_df.
}
\keyword{datasets}
