% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{preprocess_minimal}
\alias{preprocess_minimal}
\title{The result of rifi_preprocess for artificial example data
A SummarizedExperiment containing the output from rifi_preprocess}
\format{
An object of class \code{RangedSummarizedExperiment} with 4 rows and 33 columns.
}
\source{
\url{https://github.com/CyanolabFreiburg/rifi}
}
\usage{
data(preprocess_minimal)
}
\description{
The result of rifi_preprocess for artificial example data
A SummarizedExperiment containing the output from rifi_preprocess
}
\keyword{datasets}
