% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{preprocess_synechocystis_6803}
\alias{preprocess_synechocystis_6803}
\title{The result of rifi_preprocess for Synechocystis 6803 example data is a
A SummarizedExperiment containing the output of rifi_preprocess as an
extention to rowRanges}
\format{
A SummarizedExperiment:
\describe{
\item{inp:}{the SummarizedExperiment:
\describe{
\item{ID:}{The bin/probe specific ID}
\item{position:}{The bin/probe specific position}
\item{strand:}{The bin/probe specific strand}
\item{intensity:}{The relative intensity at time point 0}
\item{probe_TI:}{An internal value to determine which fitting model
is applied}
\item{flag:}{Information on which fitting model is applied}
\item{postion_segment:}{The position based segment}
}
}
\item{fit_obj_TI:}{the fit object for the TI fit:
\describe{
\item{0:}{relative intensities at 0 min}
\item{2:}{relative intensities at 2 min}
\item{4:}{relative intensities at 4 min}
\item{8:}{relative intensities at 8 min}
\item{16:}{relative intensities at 16 min}
\item{32:}{relative intensities at 32 min}
\item{64:}{relative intensities at 64 min}
\item{ID:}{The bin/probe specific ID}
\item{position:}{The bin/probe specific position}
\item{filtration:}{indicator wether the replicate is filtered or not}
}
}
}
}
\source{
\url{https://github.com/CyanolabFreiburg/rifi}
}
\usage{
data(preprocess_synechocystis_6803)
}
\description{
The result of rifi_preprocess for Synechocystis 6803 example data is a
A SummarizedExperiment containing the output of rifi_preprocess as an
extention to rowRanges
}
\keyword{datasets}
