% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{res_minimal}
\alias{res_minimal}
\title{The result of event_dataframe for E.coli artificial example.
A data frame combining the processed genome annotation and a
SummarizedExperiment data from rifi_stats. The dataframe is}
\format{
A list with 2 items:
\describe{
\item{region:}{the region from the gff file}
\item{gene:}{String, gene annotation covering the fragments}
\item{locus_tag:}{String, locus_tag annotation covering the fragments}
\item{strand:}{the strand of the annotation}
\item{TU:}{The overarching transcription unit}
\item{position:}{The bin/probe specific position}
\item{FC_fragment_intensity:}{String, fragments involved in fold change
between 2 intensity fragments}
\item{FC_intensity:}{Integer, the fold change value of 2 intensity fragments}
\item{p_value_intensity:}{p_value of the fold change of intensity fragments}
\item{FC_fragment_HL:}{Integer, the fold change value of 2 intensity fragments}
\item{FC_HL:}{Integer, the fold change value of 2 HL fragments}
\item{p_value_HL:}{p_value of the fold change of HL fragments}
\item{FC_HL_intensity_fragment:}{fragments involved on ratio of fold change between 2 half-life fragments and fold change between 2 intensity fragments}
\item{FC_HL_intensity:}{ratio of fold change between 2 half-life fragments and fold change between 2 intensity fragments}
\item{FC_HL_adapted:}{Integer, the fold change of half-life/ fold change
of intensity, position of the half-life fragment is adapted to intensity
fragment}
\item{p_value_Manova:}{p_value of the variance between two fold-changes, HL and intensity}
\item{synthesis_ratio:}{Integer, the value correspomding to synthesis rate}
\item{synthesis_ratio_event:}{String, the event assigned by synthesis rate either Termination or iTSS}
\item{pausing_site:}{presence of pausing site indicated by +/-}
\item{iTSS_I:}{presence of iTSS_I indicated by +/-}
\item{event_ps_itss_p_value_Ttest:}{p_value of pausing site or iTSS_I}
\item{ps_ts_fragment:}{The fragments involved in pausing site or iTSS_I}
\item{event_position:}{Integer, the position middle between 2 fragments with an event}
\item{event_duration:}{Integer, the duration between two delay fragments}
\item{delay_frg_slope:}{the slope value of the respective delay fragment}
\item{p_value_slope:}{p_value of the slope}
\item{delay:}
\item{half_life:}
\item{intensity:}
}
}
\source{
\url{https://github.com/CyanolabFreiburg/rifi}
}
\usage{
data(res_minimal)
}
\description{
The result of event_dataframe for E.coli artificial example.
A data frame combining the processed genome annotation and a
SummarizedExperiment data from rifi_stats. The dataframe is
}
\keyword{datasets}
