% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rifi_fragmentation.r
\name{rifi_fragmentation}
\alias{rifi_fragmentation}
\title{rifi_fragmentation: conveniently wraps all fragmentation steps}
\usage{
rifi_fragmentation(
  probe,
  cores = 1,
  logbook,
  pen_delay = logbook["delay_penalty"],
  pen_out_delay = logbook["delay_outlier_penalty"],
  pen_HL = logbook["half_life_penalty"],
  pen_out_HL = logbook["half_life_outlier_penalty"],
  pen_inty = logbook["intensity_penalty"],
  pen_out_inty = logbook["intensity_outlier_penalty"],
  pen_TU = -0.75,
  pen_TI = logbook["TI_penalty"],
  pen_out_TI = logbook["TI_outlier_penalty"]
)
}
\arguments{
\item{probe}{probe data frame: the probe based data frame.}

\item{cores}{integer: the number of assigned cores for the task.}

\item{logbook}{numeric vector: the logbook vector, if it exists.}

\item{pen_delay}{numeric: an internal parameter for the dynamic programming.
Higher values result in fewer fragments. Default is the auto generated value.}

\item{pen_out_delay}{numeric: an internal parameter for the dynamic
programming. Higher values result in fewer allowed outliers. Default is the
auto generated value.}

\item{pen_HL}{numeric: an internal parameter for the dynamic programming.
Higher values result in fewer fragments. Default is the auto generated value.}

\item{pen_out_HL}{numeric: an internal parameter for the dynamic programming.
Higher values result in fewer allowed outliers. Default is the auto generated
value.}

\item{pen_inty}{numeric: an internal parameter for the dynamic programming.
Higher values result in fewer fragments. Default is the auto generated value.}

\item{pen_out_inty}{numeric: an internal parameter for the dynamic
programming. Higher values result in fewer allowed outliers. Default is
the auto generated value.}

\item{pen_TU}{numeric: an internal parameter for the dynamic programming.
Higher values result in fewer fragments. Default -0.75.}

\item{pen_TI}{numeric: an internal parameter for the dynamic programming.
Higher values result in fewer fragments. Default is the auto generated value.}

\item{pen_out_TI}{numeric: an internal parameter for the dynamic programming.
Higher values result in fewer allowed outliers. Default is the auto generated
value.}
}
\value{
the probe data frame with the columns regarding the fragmentation:
\describe{
\item{ID:}{The bin/probe specific ID}
\item{position:}{The bin/probe specific position}
\item{strand:}{The bin/probe specific strand}
\item{intensity:}{The relative intensity at time point 0}
\item{probe_TI:}{An internal value to determine which fitting model is
applied}
\item{flag:}{Information on which fitting model is applied}
\item{position_segment:}{The position based segment}
\item{delay:}{The delay value of the bin/probe}
\item{half_life:}{The half-life of the bin/probe}
\item{TI_termination_factor:}{The termination factor of the bin/probe}
\item{delay_fragment:}{The delay fragment the bin belongs to}
\item{velocity_fragment:}{The velocity value of the respective delay
fragment}
\item{intercept:}{The vintercept of fit through the respective delay
fragment}
\item{slope:}{The slope of the fit through the respective delay fragment}
\item{HL_fragment:}{The half-life fragment the bin belongs to}
\item{HL_mean_fragment:}{The mean half-life value of the respective
half-life fragment}
\item{intensity_fragment:}{The intensity fragment the bin belongs to}
\item{intensity_mean_fragment:}{The mean intensity value of the respective
intensity fragment}
\item{TU:}{The overarching transcription unit}
\item{TI_termination_fragment:}{The TI fragment the bin belongs to}
\item{TI_mean_termination_factor:}{The mean termination factor of the
respective TI fragment}
\item{seg_ID:}{The combined ID of the fragment}
}
}
\description{
rifi_fragmentation wraps the following functions:
\enumerate{
\item fragment_delay
\item fragment_HL
\item fragment_inty
\item TUgether
\item fragment_TI
}
}
\examples{
data(fit_minimal)
data(penalties_minimal)
rifi_fragmentation(probe = fit_minimal, cores = 2,
logbook = penalties_minimal)

}
\seealso{
\code{fragment_delay}

\code{fragment_HL}

\code{fragment_inty}

\code{TUgether}

\code{fragment_TI}
}
