% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rifi_summary.r
\name{rifi_summary}
\alias{rifi_summary}
\title{rifi_summary: conveniently wraps and summarize all rifi outputs.
Wraps the functions: event_dataframe, dataframe_summary,
dataframe_summary_events, dataframe_summary_events_HL_int,
dataframe_summary_events_ps_itss, dataframe_summary_events_velocity and
dataframe_summary_TI.}
\usage{
rifi_summary(inp, data_annotation = metadata(inp)$annot[[1]])
}
\arguments{
\item{inp}{SummarizedExperiment: the input data frame with correct format.}

\item{data_annotation}{dataframe: gff3 dataframe after processing.}
}
\value{
WIP
}
\description{
rifi_summary: conveniently wraps and summarize all rifi outputs.
Wraps the functions: event_dataframe, dataframe_summary,
dataframe_summary_events, dataframe_summary_events_HL_int,
dataframe_summary_events_ps_itss, dataframe_summary_events_velocity and
dataframe_summary_TI.
}
\examples{
data(stats_minimal)
rifi_summary(inp = stats_minimal, data_annotation = 
metadata(inp)$annot[[1]])

}
\seealso{
\code{event_dataframe}

\code{dataframe_summary}

\code{dataframe_summary_events}

\code{dataframe_summary_events_HL_int}

\code{dataframe_summary_events_ps_itss}

\code{dataframe_summary_events_velocity}

\code{dataframe_summary_TI}
}
