% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rifi_summary.r
\name{rifi_summary}
\alias{rifi_summary}
\title{\subsection{=========================================================================
rifi_summary}{

rifi_summary wraps conveniently and summarize all rifi outputs
}}
\usage{
rifi_summary(inp, data_annotation = metadata(inp)$annot[[1]])
}
\arguments{
\item{inp}{SummarizedExperiment: the input data frame with correct format.}

\item{data_annotation}{dataframe: gff3 dataframe after processing.}
}
\value{
WIP
}
\description{
rifi_summary wraps the functions:
\enumerate{
\item event_dataframe
\item dataframe_summary
\item dataframe_summary_events
\item dataframe_summary_events_HL_int
\item dataframe_summary_events_ps_itss
\item dataframe_summary_events_velocity
\item dataframe_summary_TI
}
}
\examples{
data(stats_minimal)
if(!require(SummarizedExperiment)){
suppressPackageStartupMessages(library(SummarizedExperiment))
}
rifi_summary(inp = stats_minimal, data_annotation = 
metadata(stats_minimal)$annot[[1]])

}
\seealso{
\code{event_dataframe}

\code{dataframe_summary}

\code{dataframe_summary_events}

\code{dataframe_summary_events_HL_int}

\code{dataframe_summary_events_ps_itss}

\code{dataframe_summary_events_velocity}

\code{dataframe_summary_TI}
}
