% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rifi_wrapper.r
\name{rifi_wrapper}
\alias{rifi_wrapper}
\title{rifi_wrapper: conveniently wraps all functions included on rifi workflow.
Wraps the functions: check_input, rifi_preprocess, rifi_fit, rifi_penalties,
rifi_fragmentation, rifi_stats, rifi_summary and rifi_visualization.}
\usage{
rifi_wrapper(inp, cores, gff, bg, restr)
}
\arguments{
\item{inp}{data frame: the input data frame with correct format.}

\item{cores}{integer: the number of assigned cores for the task.}

\item{gff}{path: path to an annotation file in gff format}

\item{bg}{numeric: threshold over which the last time point has to be to be
fitted with the above background mode.}

\item{restr}{numeric: a parameter that restricts the freedom of the fit to
avoid wrong TI-term_factors, ranges from 0 to 0.2}
}
\value{
All intermediate objects
}
\description{
rifi_wrapper: conveniently wraps all functions included on rifi workflow.
Wraps the functions: check_input, rifi_preprocess, rifi_fit, rifi_penalties,
rifi_fragmentation, rifi_stats, rifi_summary and rifi_visualization.
}
\seealso{
\code{check_input}

\code{rifi_preprocess}
}
