% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_pos.r
\name{segment_pos}
\alias{segment_pos}
\title{segment_pos adds the column "position_segment" to the rowRanges.
To reduce run time, the data is divided by regions of no expression larger
than "dist" nucleotides.}
\usage{
segment_pos(inp, dista = 300)
}
\arguments{
\item{inp}{SummarizedExperiment: the input.}

\item{dista}{integer: the amount of nucleotides defining the gap. Default
is 300.}
}
\value{
the SummarizedExperiment object: with position_segment added to the
rowRanges.
}
\description{
segment_pos adds the column "position_segment" to the rowRanges.
To reduce run time, the data is divided by regions of no expression larger
than "dist" nucleotides.
}
\examples{
data(preprocess_minimal)
segment_pos(inp = preprocess_minimal, dista = 300)

}
