% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{summary_e_coli}
\alias{summary_e_coli}
\title{The result of rifi_summary for E.coli example data
A SummarizedExperiment containing the output of rifi_stats as an
extention to rowRanges}
\format{
The rowRanges of SummarizedExperiment:
\describe{
\item{bin_df:}{all information regarding bins:
\describe{
\item{ID:}{}
\item{feature_type:}{}
\item{gene:}{}
\item{locus_tag:}{}
\item{position:}{}
\item{strand:}{The bin/probe specific strand}
\item{segment:}{The segment the bin/probe belongs to}
\item{TU:}{The overarching transcription unit}
\item{delay_fragment:}{The delay fragment the bin/probe belongs to}
\item{delay:}{The delay of the bin/probe}
\item{HL_fragment:}{The half-life fragment the bin/probe belongs to}
\item{half_life:}{The half-life of the bin/probe}
\item{intensity_fragment:}{The intensity fragment the bin/probe belongs
to}
\item{intensity:}{The relative intensity at time point 0}
\item{flag:}{The flag of the bin/probe(TI, PDD)}
\item{TI_termination_factor:}{The TI_termination_factor of the bin/probe
(in case TI is detected)}
}
}
\item{frag_df:}{all information regarding fragments:
\describe{
\item{feature_type:}{}
\item{gene:}{}
\item{locus_tag:}{}
\item{first_position_frg:}{The first position of the fragment on the
genome}
\item{last_position_frg:}{The last position of the fragment on the
genome}
\item{strand:}{The bin/probe specific strand}
\item{TU:}{The overarching transcription unit}
\item{segment:}{The segment the fragment belongs to}
\item{delay_fragment:}{The delay fragment of the fragment}
\item{HL_fragment:}{The half-life fragment of the fragment}
\item{half_life:}{The half-life mean of the fragment}
\item{HL_SD:}{The half-life standard deviation of the fragment}
\item{HL_SE:}{The half-life standard error of the fragment}
\item{intensity_fragment:}{The intensity_fragment of the fragment}
\item{intensity:}{The relative intensity at time point 0}
\item{intensity_SD:}{The intensity standard deviation of the fragment}
\item{intensity_SE:}{The intensity standard error of the fragment}
\item{velocity:}{The velocity value of the respective delay fragment}
}
}
\item{event_df:}{all information regarding events:
\describe{
\item{event:}{}
\item{p_value:}{}
\item{p_adjusted:}{}
\item{FC_HL:}{Fold change of half-life}
\item{FC_intensity:}{Fold change of intensity}
\item{FC_HL_adapted:}{Fold change of half-life/ fold change of intensity,
position of the half-life fragment is adapted to intensity fragment}
\item{FC_HL_FC_intensity:}{Fold change of half-life/ fold change of
intensity}
\item{event_position:}{}
\item{velocity_ratio:}{}
\item{feature_type:}{}
\item{gene:}{}
\item{locus_tag:}{}
\item{strand:}{The bin/probe specific strand}
\item{TU:}{The overarching transcription unit}
\item{segment_1:}{}
\item{segment_2:}{}
\item{event_duration:}{}
\item{gap_fragments:}{}
\item{features:}{}
}
}
\item{events_HL_int_df:}{all information regarding events related to
half-life and intensity:
\describe{
\item{event:}{}
\item{p_value:}{}
\item{p_adjusted:}{}
\item{FC_HL:}{}
\item{FC_intensity:}{}
\item{FC_HL_adapted:}{Fold change of half-life/ fold change of intensity,
position of the half-life fragment is adapted to intensity fragment}
\item{FC_HL_FC_intensity:}{Fold change of half-life/ fold change of
intensity}
\item{event_position:}{}
\item{feature_type:}{}
\item{gene:}{}
\item{locus_tag:}{}
\item{strand:}{The bin/probe specific strand}
\item{TU:}{The overarching transcription unit}
\item{segment_1:}{}
\item{segment_2:}{}
\item{event_duration:}{}
\item{gap_fragments:}{}
\item{features:}{}
}
}
\item{events_ps_itss_df:}{all information regarding events related to
pausing sites and iTSS_I:
\describe{
\item{event:}{}
\item{p_value:}{}
\item{p_adjusted:}{}
\item{event_position:}{}
\item{velocity_ratio:}{}
\item{FC_HL_adapted:}{}
\item{feature_type:}{}
\item{gene:}{}
\item{locus_tag:}{}
\item{strand:}{The bin/probe specific strand}
\item{TU:}{The overarching transcription unit}
\item{segment_1:}{}
\item{segment_2:}{}
\item{event_duration:}{}
\item{gap_fragments:}{}
\item{features:}{}
}
}
\item{events_velocity_df:}{all information regarding events related to
velocity:
\describe{
\item{event:}{}
\item{p_value:}{}
\item{p_adjusted:}{}
\item{event_position:}{}
\item{velocity_ratio:}{}
\item{feature_type:}{}
\item{gene:}{}
\item{locus_tag:}{}
\item{strand:}{The bin/probe specific strand}
\item{TU:}{The overarching transcription unit}
\item{segment_1:}{}
\item{segment_2:}{}
\item{event_duration:}{}
\item{gap_fragments:}{}
\item{features:}{}
}
}
\item{TI_df:}{all information regarding TI:
\describe{
\item{event:}{}
\item{TI_fragment:}{}
\item{TI_termination_factor:}{}
\item{p_value:}{}
\item{p_adjusted:}{}
\item{feature_type:}{}
\item{gene:}{}
\item{locus_tag:}{}
\item{strand:}{The bin/probe specific strand}
\item{TU:}{The overarching transcription unit}
\item{features:}{}
\item{event_position:}{}
\item{position_1:}{}
\item{position_2:}{}
}
}
}
}
\source{
\url{https://github.com/CyanolabFreiburg/rifi}
}
\usage{
data(summary_e_coli)
}
\description{
The result of rifi_summary for E.coli example data
A SummarizedExperiment containing the output of rifi_stats as an
extention to rowRanges
}
\keyword{datasets}
