% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{summary_synechocystis_6803}
\alias{summary_synechocystis_6803}
\title{The result of rifi_summary for Synechocystis 6803 example data
A list containing the output from rifi_summary, including the fragment
based data frame, bin based data frame, event data frame and the TI
dataframe.}
\format{
A list of 4 data frames with 3000 rows and 11 variables, 297
rows and 11 variables, 486 rows and 18 variables, and 10 rows and 14
variables:
\describe{
\item{bin_df:}{all information regarding bins:
\describe{
\item{ID:}
\item{feature_type:}{String, region annotation covering the fragments}
\item{gene:}{String, gene annotation covering the fragments}
\item{locus_tag:}{String, locus_tag annotation covering the fragments}
\item{position:}
\item{strand:}{The bin/probe specific strand}
\item{segment:}{The segment the bin/probe belongs to}
\item{TU:}{The overarching transcription unit}
\item{delay_fragment:}{The delay fragment the bin/probe belongs to}
\item{delay:}{The delay of the bin/probe}
\item{HL_fragment:}{The half-life fragment the bin/probe belongs to}
\item{half_life:}{The half-life of the bin/probe}
\item{intensity_fragment:}{The intensity fragment the bin/probe belongs
to}
\item{intensity:}{The relative intensity at time point 0}
\item{flag:}{The flag of the bin/probe(TI, PDD)}
\item{TI_termination_factor:}{The TI_termination_factor of the bin/probe
(in case TI is detected)}
}
}
\item{frag_df:}{all information regarding fragments:
\describe{
\item{feature_type:}{String, region annotation covering the fragments}
\item{gene:}{String, gene annotation covering the fragments}
\item{locus_tag:}{String, locus_tag annotation covering the fragments}
\item{first_position_frg:}{The first position of the fragment on the
genome}
\item{last_position_frg:}{The last position of the fragment on the
genome}
\item{strand:}{The bin/probe specific strand}
\item{TU:}{The overarching transcription unit}
\item{segment:}{The segment the fragment belongs to}
\item{delay_fragment:}{The delay fragment of the fragment}
\item{HL_fragment:}{The half-life fragment of the fragment}
\item{half_life:}{The half-life mean of the fragment}
\item{HL_SD:}{The half-life standard deviation of the fragment}
\item{HL_SE:}{The half-life standard error of the fragment}
\item{intensity_fragment:}{The intensity_fragment of the fragment}
\item{intensity:}{The relative intensity at time point 0}
\item{intensity_SD:}{The intensity standard deviation of the fragment}
\item{intensity_SE:}{The intensity standard error of the fragment}
\item{velocity:}{The velocity value of the respective delay fragment}
}
}
\item{event_df:}{all information regarding events:
\describe{
\item{event:}{String, event type}
\item{p_value:}{Integer, p_value of the event}
\item{p_adjusted:}{Integer, p_value adjusted}
\item{FC_HL:}{Integer, the fold change value of 2 HL fragments}
\item{FC_intensity:}{Fold change of intensity}
\item{FC_HL_adapted:}{Integer, the fold change of half-life/ fold change
of intensity, position of the half-life fragment is adapted to intensity
fragment}
\item{FC_HL_FC_intensity:}{Fold change of half-life/ fold change of
intensity}
\item{event_position:}{Integer, the position middle between 2 fragments
with an event}
\item{velocity_ratio:}{Integer, ratio of velocity between 2 delay fragments}
\item{feature_type:}{String, region annotation covering the fragments}
\item{gene:}{String, gene annotation covering the fragments}
\item{locus_tag:}{String, locus_tag annotation covering the fragments}
\item{strand:}{The bin/probe specific strand}
\item{TU:}{The overarching transcription unit}
\item{segment_1:}{String, the first fragment of the two of fragments
subjected to analysis}
\item{segment_2:}{String, the second fragment of the two of fragments
subjected to analysis}
\item{event_duration:}{Integer, the duration between two delay fragments}
\item{gap_fragments:}{Integer, the distance between two delay fragments}
\item{features:}{Integer, number of fragements involved on the event}
}
}
\item{events_HL_int_df:}{all information regarding events related to
half-life and intensity:
\describe{
\item{event:}{String, event type}
\item{p_value:}{Integer, p_value of the event}
\item{p_adjusted:}{Integer, p_value adjusted}
\item{FC_HL:}{Integer, the fold change value of 2 HL fragments}
\item{FC_intensity:}{Integer, the fold change value of 2 intensity
fragments}
\item{FC_HL_adapted:}{Integer, the fold change of half-life/ fold change
of intensity,
position of the half-life fragment is adapted to intensity fragment}
\item{FC_HL_FC_intensity:}{Fold change of half-life/ fold change of
intensity}
\item{event_position:}{Integer, the position middle between 2 fragments
with an event}
\item{feature_type:}{String, region annotation covering the fragments}
\item{gene:}{String, gene annotation covering the fragments}
\item{locus_tag:}{String, locus_tag annotation covering the fragments}
\item{strand:}{The bin/probe specific strand}
\item{TU:}{The overarching transcription unit}
\item{segment_1:}{String, the first fragment of the two of fragments
subjected to analysis}
\item{segment_2:}{String, the second fragment of the two of fragments
subjected to analysis}
\item{event_duration:}{Integer, the duration between two delay fragments}
\item{gap_fragments:}{Integer, the distance between two delay fragments}
\item{features:}{Integer, number of fragements involved on the event}
}
}
\item{events_ps_itss_df:}{all information regarding events related to
pausing sites and iTSS_I:
\describe{
\item{event:}{String, event type}
\item{p_value:}{Integer, p_value of the event}
\item{p_adjusted:}{Integer, p_value adjusted}
\item{event_position:}{Integer, the position middle between 2 fragments
with an event}
\item{velocity_ratio:}{Integer, ratio of velocity between 2 delay fragments}
\item{FC_HL_adapted:}{Integer, the fold change of half-life/ fold change
of intensity,
position of the half-life fragment is adapted to intensity fragment}
\item{feature_type:}{String, region annotation covering the fragments}
\item{gene:}{String, gene annotation covering the fragments}
\item{locus_tag:}{String, locus_tag annotation covering the fragments}
\item{strand:}{The bin/probe specific strand}
\item{TU:}{The overarching transcription unit}
\item{segment_1:}{String, the first fragment of the two of fragments
subjected to analysis}
\item{segment_2:}{String, the second fragment of the two of fragments
subjected to analysis}
\item{event_duration:}{Integer, the duration between two delay fragments}
\item{gap_fragments:}{Integer, the distance between two delay fragments}
\item{features:}{Integer, number of fragements involved on the event}
}
}
\item{events_velocity_df:}{all information regarding events related to
velocity:
\describe{
\item{event:}{String, event type}
\item{p_value:}{Integer, p_value of the event}
\item{p_adjusted:}{Integer, p_value adjusted}
\item{event_position:}{Integer, the position middle between 2 fragments
with an event}
\item{velocity_ratio:}{Integer, ratio of velocity between 2 delay
fragments}
\item{feature_type:}{String, region annotation covering the fragments}
\item{gene:}{String, gene annotation covering the fragments}
\item{locus_tag:}{String, locus_tag annotation covering the fragments}
\item{strand:}{The bin/probe specific strand}
\item{TU:}{The overarching transcription unit}
\item{segment_1:}{String, the first fragment of the two of fragments
subjected to analysis}
\item{segment_2:}{String, the second fragment of the two of fragments
subjected to analysis}
\item{event_duration:}{Integer, the duration between two delay fragments}
\item{gap_fragments:}{Integer, the distance between two delay fragments}
\item{features:}{Integer, number of fragements involved on the event}
}
}
\item{TI_df:}{all information regarding TI:
\describe{
\item{event:}{String, event type}
\item{TI_fragment:}
\item{TI_termination_factor:}
\item{p_value:}{Integer, p_value of the event}
\item{p_adjusted:}{Integer, p_value adjusted}
\item{feature_type:}{String, region annotation covering the fragments}
\item{gene:}{String, gene annotation covering the fragments}
\item{locus_tag:}{String, locus_tag annotation covering the fragments}
\item{strand:}{The bin/probe specific strand}
\item{TU:}{The overarching transcription unit}
\item{features:}{Integer, number of fragements involved on the event}
\item{event_position:}{Integer, the position middle between 2 fragments with an event}
\item{position_1:}
\item{position_2:}
}
}
}
}
\source{
\url{https://github.com/CyanolabFreiburg/rifi}
}
\usage{
data(summary_synechocystis_6803)
}
\description{
The result of rifi_summary for Synechocystis 6803 example data
A list containing the output from rifi_summary, including the fragment
based data frame, bin based data frame, event data frame and the TI
dataframe.
}
\keyword{datasets}
