% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TUgether.r
\name{TUgether}
\alias{TUgether}
\title{TUgether: combines delay fragments into TUs.}
\usage{
TUgether(inp, cores = 1, pen = -0.75)
}
\arguments{
\item{inp}{SummarizedExperiment: the input data frame with correct format.}

\item{cores}{cores: integer: the number of assigned cores for the task.}

\item{pen}{numeric: an internal parameter for the dynamic programming.
Higher values result in fewer fragments. Default -0.75.}
}
\value{
the SummarizedExperiment with the columns regarding the TU:
\describe{
\item{ID:}{The bin/inp specific ID}
\item{position:}{The bin/inp specific position}
\item{position_segment:}{The position based segment}
\item{delay_fragment:}{The delay fragment the bin belongs to}
\item{intercept:}{The vintercept of fit through the respective delay
fragment}
\item{slope:}{The slope of the fit through the respective delay fragment}
\item{TU:}{The overarching transcription unit}
}
}
\description{
TUgether combines delay fragments into TUs.
The column "TU" is added.
}
\details{
TUgether combines delay fragments into TUs. It uses score fun_increasing
on the start and end points of delay_fragments.
The function used is:
.score_fun_increasing
The input is the SummarizedExperiment object.
pen is the penalty for new fragments in the dynamic programming. Since high
scores are aimed, pen is negative.
}
\examples{
data(fragmentation_minimal)
TUgether(inp = fragmentation_minimal, cores = 2, pen = -0.75)

}
